refactor: refactored TestClasses in correct packages
