+++ src/main/java/at/aau/se2/handler/game/subhandlers/SwitchCardPlayerHandler.java
@@ -42,7 +42,7 @@ public class SwitchCardPlayerHandler implements ActionHandler {
                 if(p.getUsername().equals(switchUsername)){
                     p.getCards().add(currentCard);
                     WebSocketSession switchSession= p.getSession();
-                    switchSession.sendMessage(new TextMessage(convertToJSONrequest(currentCard, UtilityMethods.findusernameofPlayer(session))));
+                    switchSession.sendMessage(new TextMessage(convertToJSONrequest(currentCard, UtilityMethods.findUsernameOfPlayer(session))));
                     break;
                 }
             }
